requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
awslimitchecker==12.0.0
blinker==1.6.2
boto3==1.28.11
botocore==1.31.11
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
Flask-RESTful==0.3.10
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
multidict==6.0.4
Pillow==10.0.0
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
s3transfer==0.6.1
six==1.16.0
smmap==5.0.0
termcolor==2.3.0
tomli==2.0.1
urllib3==1.26.16
versionfinder==1.1.1
watchdog==3.0.0
Werkzeug==2.3.6
yarl==1.9.2
